requirements.txt
boto3==1.34.0
pandas==2.2.3
pyarrow==18.1.0
pytest==8.3.5
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.10.6
mlflow==2.19.0
